JS: scroll.js 주석 완료
